bugfix: allow variables in matcher expectations
Fix a bug where the 'should' DSL function evaluates variables in the
sender context to be evaluated in the actual value, rendering them
unknown.

The matcher message now is evaluated by 'should' in the context of the
sender and then matched on the actual value.

The OperatorMatcher describes the expectation nicely in case of
early-evaluation (i.e. cached arguments).